feat: 新增basket、folder表
