+++ src-tauri/Cargo.toml
@@ -11,13 +11,13 @@ edition = "2021"
 tauri-build = { version = "1", features = [] }
 
 [dependencies]
-tauri = { version = "1", features = ["api-all"] }
+tauri = { version = "1", features = ["api-all", "linux-protocol-headers"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 dotenv = "0.15"
 sqlx = { version = "0.7.4", features = ["runtime-tokio", "sqlite"] }
 tokio = { version = "1.20.0", features = ["full"] }
-once_cell = {version = "1"}
+once_cell = { version = "1" }
 time = "0.3.34"
 image = "0.25.1"
 chrono = "0.4.37"
