+++ src-tauri/src/db/entity/mod.rs
@@ -1,0 +1,4 @@
+pub mod basket;
+pub mod folder;
+pub mod metadata;
+pub mod task;
