+++ src-tauri/src/file/mod.rs
@@ -1,8 +1,6 @@
 pub mod image_scanner;
-pub mod metadata;
 pub mod model_scanner;
 pub mod scan;
 pub mod video_scanner;
 pub mod raw_scanner;
-pub mod task;
 pub mod psd_scanner;
