+++ src-tauri/src/file/scan.rs
@@ -3,9 +3,11 @@ use std::time::Instant;
 
 use tokio::sync::mpsc::{Receiver, Sender};
 
+use crate::db::entity::basket::{Basket, BasketData};
+use crate::db::entity::folder::Folder;
+use crate::db::entity::metadata::Metadata;
+use crate::db::entity::task::{Task, TaskStatus};
 use crate::db::sqlite::Session;
-use crate::file::metadata::Metadata;
-use crate::file::task::{Task, TaskStatus};
 use crate::util::error::ErrorHandle;
 use crate::util::snowflake::id_str;
 use crate::{debug, info, Result};
@@ -34,6 +36,10 @@ pub struct ScanJob {
     pub file_count: usize,
     pub task_count: usize,
     pub scan_count: usize,
+    pub parent_folder_id: i64,
+    pub folder_list: Vec<Folder>,
+    pub basket_name: String,
+    pub directories: Vec<String>,
 }
 
 impl ScanJob {
@@ -46,6 +52,10 @@ impl ScanJob {
             file_count: 0,
             task_count: 0,
             scan_count: 0,
+            parent_folder_id: 0,
+            folder_list: Vec::new(),
+            basket_name: String::new(),
+            directories: Vec::new(),
         }
     }
 
@@ -57,6 +67,11 @@ impl ScanJob {
         self.load_dir(directories).await;
         self.load_task().await;
         self.run_scanner().await;
+
+        let mut session = Session::new("./db/main.db");
+        session.connect().await;
+        self.save_folder(&session).await;
+        self.save_basket(&session).await;
     }
 
     pub async fn load_dir(&mut self, directories: Vec<String>) {
@@ -83,8 +98,12 @@ impl ScanJob {
 
     fn load_file_list(&mut self, path: &Path) -> Result<()> {
         if path.is_dir() && !is_hidden(path) {
+            let folder = Folder::new(&path, self.parent_folder_id);
+            let folder_id = folder.id;
+            self.folder_list.push(folder);
             let read = path.read_dir()?;
             for entry in read.flatten() {
+                self.parent_folder_id = folder_id;
                 self.load_file_list(entry.path().as_path())?
             }
         } else if path.is_file() && self.is_support(path) {
@@ -168,8 +187,10 @@ impl ScanJob {
         }
     }
 
-    pub fn run_async(mut self, directories: Vec<String>) {
-        tokio::spawn(async move { self.run(directories).await });
+    pub fn run_async(mut self, basket: BasketData) {
+        self.basket_name = basket.name;
+        self.directories = basket.directories;
+        tokio::spawn(async move { self.run(self.directories.clone()).await });
     }
 
     pub fn monitor_async(&self, mut rx: Receiver<ScanMsg>) {
@@ -186,6 +207,49 @@ impl ScanJob {
             info!("<scan:{}> 扫描结束", id);
         });
     }
+
+    pub async fn save_folder(&mut self, session: &Session) {
+        let start = Instant::now();
+        for folder in self.folder_list.iter_mut() {
+            if !folder.exist(&session).await {
+                folder.save(&session).await;
+            } else {
+                if let Some(parent) = get_path_parent(&folder.path) {
+                    if let Some(parent) = Folder::get_by_path(&session, parent).await {
+                        if let Some(current) =
+                            Folder::get_by_path(&session, folder.path.clone()).await
+                        {
+                            if current.pid == 0 {
+                                folder.id = current.id;
+                                folder.pid = parent.id;
+                                folder.update(&session).await;
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        info!(
+            "<scan:{}> 保存{}个文件夹,代码运行时间为{:?}秒",
+            self.id,
+            self.folder_list.len(),
+            (Instant::now() - start).as_secs()
+        );
+    }
+
+    pub async fn save_basket(&mut self, session: &Session) {
+        let start = Instant::now();
+        let basket = Basket::new(self.basket_name.clone());
+        if !basket.exist(&session).await {
+            basket.save(&session).await;
+        }
+        basket.save_folder(&self.directories, &session).await;
+        info!(
+            "<scan:{}> 保存篮子信息,代码运行时间为{:?}秒",
+            self.id,
+            (Instant::now() - start).as_secs()
+        );
+    }
 }
 
 fn is_hidden(path: &Path) -> bool {
@@ -201,4 +265,15 @@ fn is_hidden(path: &Path) -> bool {
 
 fn get_file_suffix(path: &Path) -> Option<&str> {
     path.extension()?.to_str()
+}
+
+fn get_path_parent(path: &str) -> Option<String> {
+    let path = Path::new(path);
+    let parent = path.parent();
+    if let Some(path) = parent {
+        if let Some(path) = path.to_str() {
+            return Some(path.to_string());
+        }
+    }
+    None
 }
