+++ src-tauri/src/file/video_scanner.rs
@@ -1,11 +1,11 @@
+use base64::engine::general_purpose;
+use base64::Engine;
 use std::path::Path;
 use std::process::Command;
-use base64::Engine;
-use base64::engine::general_purpose;
 
-use crate::file::metadata::Metadata;
-use crate::file::scan::{Scanner};
-use crate::file::task::{Task, TaskStatus};
+use crate::db::entity::metadata::Metadata;
+use crate::db::entity::task::{Task, TaskStatus};
+use crate::file::scan::Scanner;
 use crate::Result;
 
 pub struct VideoScanner {}
@@ -24,7 +24,7 @@ impl Scanner for VideoScanner {
         }
     }
 
-    fn scan(&self,task: &Task) -> TaskStatus {
+    fn scan(&self, task: &Task) -> TaskStatus {
         let mut status = TaskStatus::new(task.id);
         if self.is_support(task.file_suffix.as_str()) {
             let path = task.file_path.clone();
@@ -51,11 +51,20 @@ fn analyze_video_metadata(path: &Path, metadata: &mut Metadata) -> Result<()> {
 
 /// 生成缩咯图
 fn thumbnail(path: &Path) -> Result<String> {
-
     // 定义 ffmpeg 命令
     let output = Command::new("ffmpeg")
-        .args(&["-i", path.to_str().unwrap(), "\
-        -vf", "thumbnail,scale=320:-1", "-frames:v", "1", "-f", "image2pipe", "-"])
+        .args(&[
+            "-i",
+            path.to_str().unwrap(),
+            "\
+        -vf",
+            "thumbnail,scale=320:-1",
+            "-frames:v",
+            "1",
+            "-f",
+            "image2pipe",
+            "-",
+        ])
         .output()?;
     // 检查命令执行是否成功
     if !output.status.success() {
@@ -63,36 +72,36 @@ fn thumbnail(path: &Path) -> Result<String> {
     }
     let buffer = output.stdout;
 
-        // 打开视频文件
-//     let input = input(path)?;
-//     // 查找视频流
-//     let video_stream = input.streams().best(Type::Video).ok_or(ffmpeg_next::Error::StreamNotFound)?;
-//
-//     // 创建过滤器图
-//     let mut filter_graph = Graph::new();
-//
-//     // 添加输入流
-//     let mut input_stream = filter_graph.add_input_from_stream(&video_stream)?;
-//
-//     // 添加thumbnail过滤器
-//     let mut filter = filter_graph.add("thumbnail", &[])?;
-//
-//     // 设置参数
-//     filter.set("frames", "1")?; // 设置提取的帧数量
-//     filter.set("scale", "320:-1")?; // 设置从第一帧开始提取
-//
-//     // 链接过滤器和输出流
-//     let mut output_stream = filter.outputs().get_mut(0).ok_or(ffmpeg_next::Error::StreamNotFound)?;
-//     output_stream.set_format(ffmpeg_next::format::Pixel::RGB24)?;
-//
-//     // 配置过滤器图
-//     filter_graph.configure()?;
-//
-//     // 执行过滤器图
-//     filter_graph.run()?;
-//
-//     // 保存输出
-//     let mut buffer = output_stream.pull_packet()?;
+    // 打开视频文件
+    //     let input = input(path)?;
+    //     // 查找视频流
+    //     let video_stream = input.streams().best(Type::Video).ok_or(ffmpeg_next::Error::StreamNotFound)?;
+    //
+    //     // 创建过滤器图
+    //     let mut filter_graph = Graph::new();
+    //
+    //     // 添加输入流
+    //     let mut input_stream = filter_graph.add_input_from_stream(&video_stream)?;
+    //
+    //     // 添加thumbnail过滤器
+    //     let mut filter = filter_graph.add("thumbnail", &[])?;
+    //
+    //     // 设置参数
+    //     filter.set("frames", "1")?; // 设置提取的帧数量
+    //     filter.set("scale", "320:-1")?; // 设置从第一帧开始提取
+    //
+    //     // 链接过滤器和输出流
+    //     let mut output_stream = filter.outputs().get_mut(0).ok_or(ffmpeg_next::Error::StreamNotFound)?;
+    //     output_stream.set_format(ffmpeg_next::format::Pixel::RGB24)?;
+    //
+    //     // 配置过滤器图
+    //     filter_graph.configure()?;
+    //
+    //     // 执行过滤器图
+    //     filter_graph.run()?;
+    //
+    //     // 保存输出
+    //     let mut buffer = output_stream.pull_packet()?;
 
     // 将输出转换为 Base64 编码的字符串
     let base64 = general_purpose::STANDARD.encode(&buffer);
