+++ src-tauri/src/main.rs
@@ -12,7 +12,10 @@ async fn main() {
             basket::create_basket,
             basket::get_metadata,
             basket::del_metadata,
-            basket::get_metadata_by_id
+            basket::get_metadata_by_id,
+            basket::get_basket,
+            basket::del_basket,
+            basket::get_folder
         ])
         .run(tauri::generate_context!())
         .print_error();
